pyproject.toml
[tool.poetry]
name = "e-commerce"
version = "0.1.0"
description = ""
authors = ["hayaa1234 <hayalawansah@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.1"
Pillow = "^9.0.0"
django-mptt = "^0.13.4"
djangorestframework = "^3.13.1"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
